+++ src/full/Agda/Interaction/Highlighting/FromAbstract.hs
@@ -362,7 +362,7 @@ instance Hilite A.LetBinding where
   hilite = \case
       A.LetBind    _r ai x t e     -> hl ai <> hl x <> hl t <> hl e
       A.LetAxiom   _r ai x t       -> hl ai <> hl x <> hl t
-      A.LetPatBind _r p e          -> hl p  <> hl e
+      A.LetPatBind _r ai p e       -> hl ai <> hl p  <> hl e
       A.LetApply mi er x es _c dir -> hl mi <> hl er <> hl x <>
                                       hl es <> hl dir
       A.LetOpen mi x dir           -> hl mi <> hl x <> hl dir