+++ src/full/Agda/Syntax/Concrete/Definitions.hs
@@ -148,14 +148,14 @@ data DeclKind
   deriving (Eq, Show)
 
 declKind :: NiceDeclaration -> DeclKind
-declKind (FunSig r _ _ _ _ _ tc cc x _)      = LoneSigDecl r (FunName tc cc) x
+declKind (FunSig r _ _ _ _ ai tc cc x _)     = LoneSigDecl r (FunName ai tc cc) x
 declKind (NiceRecSig r _ _ _ pc uc x _ _)    = LoneSigDecl r (RecName pc uc) x
 declKind (NiceDataSig r _ _ _ pc uc x _ _)   = LoneSigDecl r (DataName pc uc) x
-declKind (FunDef _r _ _ _ tc cc x _)         = LoneDefs (FunName tc cc) [x]
+declKind (FunDef _r _ _ _ tc cc x (cl :| _)) = LoneDefs (FunName (getArgInfo cl) tc cc) [x]
 declKind (NiceDataDef _ _ _ pc uc x _par _)  = LoneDefs (DataName pc uc) [x]
 declKind (NiceUnquoteData _ _ _ pc uc x _ _) = LoneDefs (DataName pc uc) [x]
 declKind (NiceRecDef _ _ _ pc uc x _ _par _) = LoneDefs (RecName pc uc) [x]
-declKind (NiceUnquoteDef _ _ _ tc cc xs _)   = LoneDefs (FunName tc cc) xs
+declKind (NiceUnquoteDef _ _ _ tc cc xs _)   = LoneDefs (FunName empty tc cc) xs
 declKind Axiom{}                             = OtherDecl
 declKind NiceField{}                         = OtherDecl
 declKind PrimitiveFunction{}                 = OtherDecl
@@ -300,7 +300,7 @@ niceDeclarations fixs ds = do
           -- (like @x y z : A@) into several type signatures (with imprecise ranges).
           let r = getRange x
           -- register x as lone type signature, to recognize clauses later
-          x' <- addLoneSig r x $ FunName termCheck covCheck
+          x' <- addLoneSig r x $ FunName info termCheck covCheck
           return ([FunSig r PublicAccess ConcreteDef NotInstanceDef NotMacroDef info termCheck covCheck x' t] , ds)
 
         -- Should not show up: all FieldSig are part of a Field block
@@ -318,7 +318,7 @@ niceDeclarations fixs ds = do
             _ -> __IMPOSSIBLE__
           return (gs, ds)
 
-        FunClause lhs _ _ _ -> do
+        FunClause ai lhs _ _ _ -> do
           termCheck <- use terminationCheckPragma
           covCheck  <- use coverageCheckPragma
           catchall  <- popCatchallPragma
@@ -339,7 +339,7 @@ niceDeclarations fixs ds = do
               -- Subcase: The lhs is single identifier (potentially anonymous).
               -- Treat it as a function clause without a type signature.
               LHS p [] [] | Just x <- isSingleIdentifierP p -> do
-                d  <- mkFunDef (setOrigin Inserted defaultArgInfo) termCheck covCheck x Nothing $ singleton d -- fun def without type signature is relevant
+                d  <- mkFunDef (setOrigin Inserted ai) termCheck covCheck x Nothing $ singleton d -- fun def without type signature can have modality
                 return (d , ds)
               -- Subcase: The lhs is a proper pattern.
               -- This could be a let-pattern binding. Pass it on.
@@ -348,11 +348,11 @@ niceDeclarations fixs ds = do
                 return ([NiceFunClause (getRange d) PublicAccess ConcreteDef termCheck covCheck catchall d] , ds)
 
             -- case: clauses match exactly one of the sigs
-            [(x,(x',fits,rest))] -> do
+            [(x, (Arg info x', fits, rest))] -> do
                -- The x'@NoName{} is the unique version of x@NoName{}.
                removeLoneSig x
-               ds  <- expandEllipsis1 fits
-               cs  <- mkClauses1 x' ds empty
+               ds <- expandEllipsis1 fits
+               cs <- mkClauses1 info x' ds empty
                return ([FunDef (getRange fits) fits ConcreteDef NotInstanceDef termCheck covCheck x' cs] , rest)
 
             -- case: clauses match more than one sigs (ambiguity)
@@ -775,7 +775,7 @@ niceDeclarations fixs ds = do
       -> Nice [NiceDeclaration]
     mkFunDef info termCheck covCheck x mt ds0 = do
       ds <- expandEllipsis1 ds0
-      cs <- mkClauses1 x ds empty
+      cs <- mkClauses1 info x ds empty
       return [ FunSig (fuseRange x t) PublicAccess ConcreteDef NotInstanceDef NotMacroDef info termCheck covCheck x t
              , FunDef (getRange ds0) ds0 ConcreteDef NotInstanceDef termCheck covCheck x cs ]
         where
@@ -790,7 +790,7 @@ niceDeclarations fixs ds = do
     expandEllipsis (d : ds) = List1.toList <$> expandEllipsis1 (d :| ds)
 
     expandEllipsis1 :: List1 Declaration -> Nice (List1 Declaration)
-    expandEllipsis1 (d@(FunClause (LHS p _ _) _ _ _) :| ds)
+    expandEllipsis1 (d@(FunClause _ (LHS p _ _) _ _ _) :| ds)
       | hasEllipsis p = (d :|) <$> expandEllipsis ds
       | otherwise     = (d :|) <$> expand (killRange p) ds
       where
@@ -800,13 +800,13 @@ niceDeclarations fixs ds = do
           case d of
             Pragma (CatchallPragma _) -> do
                   (d :) <$> expand p ds
-            FunClause (LHS p0 eqs es) rhs wh ca -> do
+            FunClause ai (LHS p0 eqs es) rhs wh ca -> do
               case hasEllipsis' p0 of
                 ManyHoles -> declarationException $ MultipleEllipses p0
                 OneHole cxt ~(EllipsisP r Nothing) -> do
                   -- Replace the ellipsis by @p@.
                   let p1 = cxt $ EllipsisP r $ Just $ setRange r p
-                  let d' = FunClause (LHS p1 eqs es) rhs wh ca
+                  let d' = FunClause ai (LHS p1 eqs es) rhs wh ca
                   -- If we have with-expressions (es /= []) then the following
                   -- ellipses also get the additional patterns in p0.
                   (d' :) <$> expand (if null es then p else killRange p1) ds
@@ -822,39 +822,48 @@ niceDeclarations fixs ds = do
     expandEllipsis1 _ = __IMPOSSIBLE__
 
     -- Turn function clauses into nice function clauses.
-    mkClauses1 :: Name -> List1 Declaration -> Catchall -> Nice (List1 Clause)
-    mkClauses1 x ds ca = List1.fromListSafe __IMPOSSIBLE__ <$> mkClauses x (List1.toList ds) ca
+    mkClauses1 ::
+         ArgInfo
+           -- Clauses need to conform to this 'Modality'.
+      -> Name -> List1 Declaration -> Catchall -> Nice (List1 Clause)
+    mkClauses1 info x ds ca = List1.fromListSafe __IMPOSSIBLE__ <$> mkClauses info x (List1.toList ds) ca
 
-    mkClauses :: Name -> [Declaration] -> Catchall -> Nice [Clause]
-    mkClauses _ [] _ = return []
+    mkClauses ::
+         ArgInfo
+           -- Clauses need to conform to this 'Modality'.
+      -> Name -> [Declaration] -> Catchall -> Nice [Clause]
+    mkClauses _ _ [] _ = return []
 
     -- A CATCHALL pragma after the last clause is useless.
-    mkClauses _ [Pragma (CatchallPragma r)] _ = [] <$ do
+    mkClauses _ _ [Pragma (CatchallPragma r)] _ = [] <$ do
       declarationWarning $ InvalidCatchallPragma r
 
-    mkClauses x (Pragma (CatchallPragma r) : cs) catchall = do
+    mkClauses info x (Pragma (CatchallPragma r) : cs) catchall = do
       -- Warn about consecutive CATCHALL pragmas
       unless (null catchall) $ declarationWarning $ InvalidCatchallPragma r
-      mkClauses x cs (YesCatchall r)
+      mkClauses info x cs (YesCatchall r)
 
-    mkClauses x (FunClause lhs rhs wh ca : cs) catchall
-      | null (lhsWithExpr lhs) || hasEllipsis lhs  =
-      (Clause x (ca <> catchall) lhs rhs wh [] :) <$> mkClauses x cs empty   -- Will result in an error later.
+    mkClauses info x (FunClause ai lhs rhs wh ca : cs) catchall
+      | null (lhsWithExpr lhs) || hasEllipsis lhs = do
+      ai <- checkModality info ai
+      (Clause x (ca <> catchall) ai lhs rhs wh [] :) <$> mkClauses info x cs empty
 
-    mkClauses x (FunClause lhs rhs wh ca : cs) catchall = do
+    mkClauses info x (FunClause ai lhs rhs wh ca : cs) catchall = do
+      ai <- checkModality info ai
       when (null withClauses) $ declarationException $ MissingWithClauses x lhs
-      wcs <- mkClauses x withClauses empty
-      (Clause x (ca <> catchall) lhs rhs wh wcs :) <$> mkClauses x cs' empty
+      wcs <- mkClauses info x withClauses empty
+      (Clause x (ca <> catchall) ai lhs rhs wh wcs :) <$> mkClauses info x cs' empty
       where
         (withClauses, cs') = subClauses cs
 
         -- A clause is a subclause if the number of with-patterns is
         -- greater or equal to the current number of with-patterns plus the
         -- number of with arguments.
-        numWith = numberOfWithPatterns p + length (filter visible es) where LHS p _ es = lhs
+        numWith = numberOfWithPatterns p + length (filter visible es)
+          where LHS p _ es = lhs
 
         subClauses :: [Declaration] -> ([Declaration],[Declaration])
-        subClauses (c@(FunClause (LHS p0 _ _) _ _ _) : cs)
+        subClauses (c@(FunClause _ (LHS p0 _ _) _ _ _) : cs)
          | isEllipsis p0 ||
            numberOfWithPatterns p0 >= numWith = mapFst (c:) (subClauses cs)
          | otherwise                           = ([], c:cs)
@@ -863,7 +872,15 @@ niceDeclarations fixs ds = do
           (cs, cs') -> (c:cs, cs')
         subClauses [] = ([],[])
         subClauses _  = __IMPOSSIBLE__
-    mkClauses _ _ _ = __IMPOSSIBLE__
+    mkClauses _ _ _ _ = __IMPOSSIBLE__
+
+    checkModality ::
+         ArgInfo       -- Authoritative 'Modality' from type signature.
+      -> ArgInfo       -- 'Modality' from clause, has to conform.
+      -> Nice ArgInfo  -- Deviant 'Modality' conformed to authoritative source.
+    checkModality info ai
+      | (null ai || sameModality info ai) = return ai
+      | otherwise = info <$ declarationWarning (DivergentModalityInClause info ai)
 
     couldBeCallOf :: Maybe Fixity' -> Name -> Pattern -> Bool
     couldBeCallOf mFixity x p =
@@ -902,7 +919,7 @@ niceDeclarations fixs ds = do
     -- for finding clauses for a type sig in mutual blocks
     couldBeFunClauseOf :: Maybe Fixity' -> Name -> Declaration -> Bool
     couldBeFunClauseOf _ _ (Pragma (CatchallPragma{})) = True
-    couldBeFunClauseOf mFixity x (FunClause (LHS p _ _) _ _ _) =
+    couldBeFunClauseOf mFixity x (FunClause _ (LHS p _ _) _ _ _) =
        hasEllipsis p || couldBeCallOf mFixity x p
     couldBeFunClauseOf _ _ _ = False -- trace ("couldBe not (fun default)") $ False
 
@@ -1001,9 +1018,10 @@ niceDeclarations fixs ds = do
 
         addFunClauses ::
              KwRange
+               -- Range of the @interleaved mutual@ keyword.
           -> [NiceDeclaration]
           -> INice [(DeclNum, NiceDeclaration)]
-        addFunClauses r (nd@(NiceFunClause _ _ _ tc cc _ (FunClause lhs _ _ _)) : ds) = do
+        addFunClauses r (nd@(NiceFunClause _ _ _ tc cc _ (FunClause ai lhs _ _ _)) : ds) = do
           -- get the candidate functions that are in this interleaved mutual block
           ISt m checks i <- get
           let sigs = mapMaybe (\ (n, d) -> n <$ isInterleavedFun d) $ Map.toList m
@@ -1022,7 +1040,7 @@ niceDeclarations fixs ds = do
             [(n, fits0, rest)] -> do
               let (checkss, fits) = List1.unzip fits0
               ds <- lift $ expandEllipsis1 fits
-              cs <- lift $ mkClauses1 n ds empty
+              cs <- lift $ mkClauses1 ai n ds empty
               addInterleavedFun n fits {- or? ds -} cs $ Fold.fold checkss
               groupByBlocks r rest
             -- more than one candidate: fail, complaining about the ambiguity!
@@ -1033,6 +1051,7 @@ niceDeclarations fixs ds = do
 
         groupByBlocks ::
               KwRange
+                -- Range of the @interleaved mutual@ keyword.
           -> [NiceDeclaration]
           -> INice [(DeclNum, NiceDeclaration)]
         groupByBlocks _kwr []      = pure []
@@ -1461,8 +1480,8 @@ instance MakeAbstract NiceDeclaration where
     NiceOpaque r ns ds                   -> NiceOpaque r ns <$> mkAbstract kwr ds
 
 instance MakeAbstract Clause where
-  mkAbstract kwr (Clause x catchall lhs rhs wh with) = do
-    Clause x catchall lhs rhs <$> mkAbstract kwr wh <*> mkAbstract kwr with
+  mkAbstract kwr (Clause x catchall ai lhs rhs wh with) = do
+    Clause x catchall ai lhs rhs <$> mkAbstract kwr wh <*> mkAbstract kwr with
 
 -- | Contents of a @where@ clause are abstract if the parent is.
 --
@@ -1542,8 +1561,8 @@ instance MakePrivate NiceDeclaration where
       d@NiceUnquoteData{}                      -> return d
 
 instance MakePrivate Clause where
-  mkPrivate kwr o (Clause x catchall lhs rhs wh with) = do
-    Clause x catchall lhs rhs <$> mkPrivate kwr o wh <*> mkPrivate kwr o with
+  mkPrivate kwr o (Clause x catchall ai lhs rhs wh with) = do
+    Clause x catchall ai lhs rhs <$> mkPrivate kwr o wh <*> mkPrivate kwr o with
 
 instance MakePrivate WhereClause where
   mkPrivate kwr o = \case