+++ CHANGELOG.md
@@ -17,7 +17,14 @@ Warnings
 * `UselessPragma` warning instead of hard error `NeedOptionRewriting` when a
   `REWRITE` or `BUILTIN REWRITE` pragma is encountered but `--rewriting` is off.
 
-* New warning 'InvalidTacticAttribute' for misplaced `@(tactic ...)` attributes.
+* New warning `DivergentModalityInClause` when modality of a clause diverges
+  from that of the function.  Example:
+  ```agda
+  A : Set₁
+  @0 A = Set
+  ```
+
+* New warning `InvalidTacticAttribute` for misplaced `@(tactic ...)` attributes.
   This was silently accepted up to Agda 2.8.0 but raises now the new warning:
   ```agda
   postulate
@@ -40,6 +47,16 @@ Additions to the Agda syntax.
   ```
 
   See [#4275](https://github.com/agda/agda/issues/4275) for the proposal.
+
+* Modality annotations in aliases and let-bindings are now supported
+  (PR [#7990]()).
+  Example:
+  ```agda
+    split : {A B C : Set} (@0 p : A × B) (k : @0 A → @0 B → C) → C
+    split p k = let @0 (x , y) = q in k x y
+      where
+      @0 q = p
+  ```
 
 Language
 --------